Added new apis to work with the mobile app
